Managing virtual environment better
When using tmux, all the virtual environment load once, then
every time a new pane is opened, it loads again.
It is unnecessary, since all panes are forks of the initial process
that loaded it anyway, and it can take seconds to load all virtual
environment (gvm, rvm, pyenv, ...), thus making tmux slow to use.

By introducing checks for every virtual environment, we go around
this issue, making opening panes much faster

We also started running tmux by default when running a login shell.
It still possible to simply exit it if it's not wanted ()for instance
for forwarding X.